refactor: cleanup plugin -> src-plugin
