+++ src-plugin/index.ts
@@ -1,0 +1,4 @@
+export { alias } from './alias'
+export { copy } from './copy'
+export { customStart } from './custom-start'
+export { loadViteEnv } from './load-vite-env'
